New Weekly Summary of The Work and contents session with accessibility changes to files (#41)
* UPDATE to the deadline table

The deadline of the meeting today has passed and the next meeting is awaiting.

* Weekly Summary of The Work and some minor update to the deadline tables

Weekly Summary of The Work and some minor update to the deadline tables inc new header and adjustment to the vocabs as well as the adjust of the layout of the file.

* Adding table of contents and fix minor bugs

Adding table of contents and fix minor bugs of the hyperlinks.  Adding going back hyperlinks to all names to go back to the weekly summary title for accessability purpose.

* Minor bug fixes and adding going back options to contents on titles

Minor bug fixes which causes clicking the hyperlinks will always going to the contents and adding going back options to contents on titles for accessibility purpose.

* (Have been proved)Adjustment to the files and folders and more data visualisation graph by Zihan

remove duplication of the files and changing the folder names.  More data visualisation graphs have been added.  All these actions have been proved by the original authors.

* Merge logisticRegression folders into one

Merge logisticRegression folders into one and put both Ernest and Bang's code into one folder with different subfolder and names.